[2] Rename json.py to json_utils.py to avoid clashing with stdlib
Fixes #2
